Learning/Learning/requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
Pillow==11.0.0